Attempt at fixing dead code lints
